Add experimental new api at tiny-atom/future
- action packs
- where each pack is of shape { state, actions }
- all gets/splits inside actions are scoped to that pack
- actions are bound to
- exposes a new method atom.fuse(actions) for extending the action packs

This is built purely on top of the tiny-atom/core with no modifications.
It demos a very decent default approach to working with tiny-atom. Whilst
keeping tiny-atom/core available for any alternative strategies.

Most of all, this preserves the combined effect/reducer approach of
tiny-atom, it preserves the custom merge strategy and is compatible
with all of the existing batteries log/devtool/react/preact.

Finally, it's backwards compatible, you should be able to utilise this in
an existing project gradually if no custom evolve was used.