Restructure the repo with subpackages nested under src to fix distribution
The imports need to be relative to the way the final package is built
I.e. things are nested under tiny-atom/react/esm by using the package.json
module entry. This means anything inside this dir needs to do ../../core
and not ../core. This is why we move everything into src so that the hierarchy
is the same in dev and in prod.